Moved tracer to statestore
