+++ fedn/cli/run_cmd.py
@@ -192,10 +192,12 @@ def dashboard_cmd(ctx, host, port, secret_key, local_package, name, init):
     statestore_config = fedn_config['statestore']
     if statestore_config['type'] == 'MongoDB':
         statestore = MongoStateStore(
-            network_id, statestore_config['mongo_config'], fedn_config['storage'])
+            network_id, statestore_config['mongo_config'])
     else:
         print("Unsupported statestore type, exiting. ", flush=True)
         exit(-1)
+
+    statestore.set_storage_backend(fedn_config['storage'])
 
     # Enable JWT token authentication.
     if config['secret_key']:
