--- fedn/fedn/common/tracer/__init__.py
@@ -1,3 +1,0 @@
-#
-# Scaleout Systems AB
-# __author__ = 'Morgan Ekmefjord morgan@scaleout.se'
