--- fedn/fedn/common/tracer/tracer.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Tracer(ABC):
-    @abstractmethod
-    def report_status(self, msg):
-        """
-
-        :param msg:
-        """
-        pass
